backend/requirements.txt
asgiref==3.8.1
axios==0.4.0
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flake8==7.1.1
idna==3.10
iniconfig==2.0.0
lxml==5.3.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
rich==13.8.1
sqlparse==0.5.1
urllib3==2.2.3
